refacto: move service code to an internal dedicated file
