+++ internal/handlers_test.go
@@ -35,7 +35,7 @@ func TestGetMsgHandlerSuccess(t *testing.T) {
 	c := e.NewContext(req, rec)
 
 	s := &FakeSecretMsgStorer{msg: "secret"}
-	h := NewSecretHandlers(s)
+	h := newSecretHandlers(s)
 	err := h.GetMsgHandler(c)
 	if err != nil {
 		t.Fatalf("got error %v, none expected", err)
@@ -63,7 +63,7 @@ func TestGetMsgHandlerError(t *testing.T) {
 	c := e.NewContext(req, rec)
 
 	s := &FakeSecretMsgStorer{msg: "secret", err: errors.New("expired")}
-	h := NewSecretHandlers(s)
+	h := newSecretHandlers(s)
 	err := h.GetMsgHandler(c)
 	if err == nil {
 		t.Fatalf("got no error, expected one")
@@ -85,7 +85,7 @@ func TestHealthHandler(t *testing.T) {
 	rec := httptest.NewRecorder()
 	c := e.NewContext(req, rec)
 
-	err := HealthHandler(c)
+	err := healthHandler(c)
 	if err != nil {
 		t.Fatalf("error returned %v, expected nil", err)
 	}
@@ -101,7 +101,7 @@ func TestRedirectHandler(t *testing.T) {
 	rec := httptest.NewRecorder()
 	c := e.NewContext(req, rec)
 
-	err := RedirectHandler(c)
+	err := redirectHandler(c)
 	if err != nil {
 		t.Fatalf("error returned %v, expected nil", err)
 	}
